+++ src/main/java/calculator/dto/SplittedResult.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.dto;
 
 import calculator.util.IntegerConverter;
