+++ src/test/java/calculator/SplittedResultTest.java
@@ -1,7 +1,6 @@
 package calculator;
 
-import calculator.splitor.BasicSplitor;
-import calculator.splitor.Splitor;
+import calculator.dto.SplittedResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +8,6 @@ import java.util.List;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class SplittedResultTest {
