+++ src/test/java/calculator/util/IntegerConverterTest.java
@@ -1,8 +1,5 @@
 package calculator.util;
 
-import calculator.InputString;
-import calculator.splitor.BasicSplitor;
-import calculator.splitor.Splitor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +7,6 @@ import java.util.List;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class IntegerConverterTest {
     @DisplayName("문자열 리스트를 숫자 리스트로 변환한다.")
